chore: Update nltk version to fix vulnerability issue. (#5218)
## Context
Update nltk version to fix vulnerability issue. Also make nltk an
optional dependency.

https://github.com/advisories/GHSA-7p94-766c-hgjp

## Change Summary
nltk version was made >= 3.9.3, where the vulnerability has been
addressed.

## Impact
None to users. Package will use a non-vulnerable version of nltk.

---------
