+++ doc/changelog.d/5218.maintenance.md
@@ -1,0 +1,1 @@
+Update nltk version to fix vulnerability issue.
